wins on rare occasions but fails to build techlabs and reactors
It doesnt fail, FAIL. it can build techlabs and reactors but it just won't due to training marines being 90% of the actions chosen by the q table. So it never gets to be free to build techlabs and reactors :/ so beyond that its hard for it to win. Also the reward system given is not working correctly